+++ Engine/Core/CMakeLists.txt
@@ -1,6 +1,7 @@
 # created on 2021/6/2
 # author @zoloypzuo
 
+add_subdirectory(Controller)
 add_subdirectory(ECS)
 add_subdirectory(EventSystem)
 add_subdirectory(Scene)
@@ -17,6 +18,7 @@ add_subdirectory(RHI)
 add_subdirectory(UI)
 add_subdirectory(Window)
 
+set_target_properties(Controller PROPERTIES FOLDER "Core")
 set_target_properties(ECS PROPERTIES FOLDER "Core")
 set_target_properties(EventSystem PROPERTIES FOLDER "Core")
 set_target_properties(Scene PROPERTIES FOLDER "Core")
@@ -36,6 +38,7 @@ set_target_properties(Window PROPERTIES FOLDER "Core")
 add_library(Core INTERFACE)
 target_include_directories(Core INTERFACE ${CMAKE_SOURCE_DIR}/Engine)
 target_link_libraries(Core INTERFACE
+        Controller
         ECS
         EventSystem
         Scene
