+++ Engine/Core/Controller/CMakeLists.txt
@@ -1,0 +1,4 @@
+add_library(Controller CFreeLook.cpp CFreeMove.cpp)
+
+target_link_libraries(Controller PUBLIC ECS)
+target_link_libraries(Controller PRIVATE Input)
