+++ Engine/Core/Parser/CMakeLists.txt
@@ -4,3 +4,5 @@ add_library(Parser
         IniReader.cpp IniReader.h
         MeshLoader.cpp MeshLoader.h
         )
+
+target_link_libraries(Parser PRIVATE assimp::assimp)
