+++ Engine/Core/RHI/CMakeLists.txt
@@ -31,7 +31,7 @@ add_library(RHI
         MeshGen/Ingredients/torus.cpp
         MeshGen/Ingredients/torus.h
 
-        Object/Camera.cpp
+        Object/ACamera.cpp
         Object/ALight.cpp
 
         RenderSystem.cpp RenderSystem.h
