+++ Engine/Core/RHI/RenderSystem.h
@@ -9,7 +9,7 @@
 #include "Core/RHI/RenderSystem.h"
 #include "Core/RHI/RenderCommand.h"
 #include "Core/RHI/RenderPipeline.h"
-#include "Core/RHI/Object/Camera.h"
+#include "Core/RHI/Object/ACamera.h"
 #include "Core/RHI/Resource/MeshManager.h"
 
 namespace Zelo::Core::RHI {
