+++ Engine/Renderer/OpenGL/GLLoader.h
@@ -1,0 +1,9 @@
+// GLLoader.h
+// created on 2021/12/5
+// author @zoloypzuo
+#pragma once
+
+#include "ZeloPrerequisites.h"
+#include "ZeloGLPrerequisites.h"
+
+void loadGL();
